Add Specification class
Encapsulate logic of getting specification properties from config file.
We can create an instance with a certain specification or with a default one.
If we can’t find property in specified spec, we’ll look in default spec
Added a static method for getting specification names from config(used for parameterised tests)